hte: modify gpio controller match function
commit 3c447bb7ec488a2b3e710464536b29a4647047c6 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git jetson_36.3

The linux jammy kernel does not have fwnode member in the
struct gpio_chip. This gives an error when using the upstream
patch for the tegra234 HTE provider. The CL addresses that
concern and implements match function to match the of_node
instead of the fwnnode

Bug 3961133

Change-Id: I542b03e98c263e9cfca7be071cd2a048cd28d948
Signed-off-by: Dipen Patel <dipenp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/canonical/linux-jammy/+/3088882
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>
Tested-by: Gautham Srinivasan <gauthams@nvidia.com>
Reviewed-by: Gautham Srinivasan <gauthams@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>