usb: typec: ucsi: Find connector fwnode without relying on order
commit 600e2d57d6a1a5d9a87f7c77a1e5fd72f5658974 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git jetson_36.3

BugLink: https://bugs.launchpad.net/bugs/2028400

The previous code relied on the order of "connector" nodes under the
ucsi device node to determine the correct fwnode handle for a connector.
However, the order of nodes should not be relied upon when parsing device
trees, especially when using live DT overlays.

This change modifies ucsi_find_fwnode() to find the correct connector
fwnode by reading the "reg" property of each connector node and
comparing it to the connector number. If only a single connector node
exists without a "reg" property, that single node is returned.

This ensures the correct connector fwnode is found regardless of node
order in the DT.

Signed-off-by: Wayne Chang <waynec@nvidia.com>
Reviewed-by: Sing-Han Chen <singhanc@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>